list関連の処理を本体から分離した
