+++ scheme/list.go
@@ -1,0 +1,316 @@
+/*
+   Go lang 3rd study program.
+   This is prototype program mini scheme subset.
+
+   hidekuno@gmail.com
+*/
+package scheme
+
+import (
+	"reflect"
+	"strconv"
+)
+
+// map,filter,reduce
+func listFunc(lambda func(Expression, Expression, []Expression) ([]Expression, error), exp ...Expression) (Expression, error) {
+	if len(exp) != 2 {
+		return nil, NewRuntimeError("E1007", strconv.Itoa(len(exp)))
+	}
+	fn, ok := exp[0].(*Function)
+	if !ok {
+		return nil, NewRuntimeError("E1006", reflect.TypeOf(exp[0]).String())
+	}
+	l, ok := exp[1].(*List)
+	if !ok {
+		return nil, NewRuntimeError("E1005", reflect.TypeOf(exp[1]).String())
+	}
+	var vaList []Expression
+	param := make([]Expression, 1)
+	for _, param[0] = range l.Value {
+		result, err := fn.Execute(param, nil)
+		if err != nil {
+			return nil, err
+		}
+		vaList, err = lambda(result, param[0], vaList)
+		if err != nil {
+			return nil, err
+		}
+	}
+	return NewList(vaList), nil
+}
+
+// Build Global environement.
+func buildListFunc() {
+
+	// list operator
+	buildInFuncTbl["list"] = func(exp []Expression, env *SimpleEnv) (Expression, error) {
+		return EvalCalcParam(exp, env,
+			func(exp ...Expression) (Expression, error) {
+				var l []Expression
+				return NewList(append(l, exp...)), nil
+			})
+	}
+	buildInFuncTbl["null?"] = func(exp []Expression, env *SimpleEnv) (Expression, error) {
+		return EvalCalcParam(exp, env,
+			func(exp ...Expression) (Expression, error) {
+				if len(exp) != 1 {
+					return nil, NewRuntimeError("E1007", strconv.Itoa(len(exp)))
+				}
+				if l, ok := exp[0].(*List); ok {
+					return NewBoolean(0 == len(l.Value)), nil
+				} else {
+					return NewBoolean(false), nil
+				}
+			})
+	}
+	buildInFuncTbl["length"] = func(exp []Expression, env *SimpleEnv) (Expression, error) {
+		return EvalCalcParam(exp, env,
+			func(exp ...Expression) (Expression, error) {
+				if len(exp) != 1 {
+					return nil, NewRuntimeError("E1007", strconv.Itoa(len(exp)))
+				}
+				if l, ok := exp[0].(*List); ok {
+					return NewInteger(len(l.Value)), nil
+				} else {
+					return nil, NewRuntimeError("E1005", reflect.TypeOf(exp[0]).String())
+				}
+			})
+	}
+	buildInFuncTbl["car"] = func(exp []Expression, env *SimpleEnv) (Expression, error) {
+		return EvalCalcParam(exp, env,
+			func(exp ...Expression) (Expression, error) {
+				if len(exp) != 1 {
+					return nil, NewRuntimeError("E1007", strconv.Itoa(len(exp)))
+				}
+				if l, ok := exp[0].(*List); ok {
+					if len(l.Value) <= 0 {
+						return nil, NewRuntimeError("E1011", strconv.Itoa(len(l.Value)))
+					}
+					return l.Value[0], nil
+				} else if p, ok := exp[0].(*Pair); ok {
+					return p.Car, nil
+				} else {
+					return nil, NewRuntimeError("E1005", reflect.TypeOf(exp[0]).String())
+				}
+			})
+	}
+	buildInFuncTbl["cdr"] = func(exp []Expression, env *SimpleEnv) (Expression, error) {
+		return EvalCalcParam(exp, env,
+			func(exp ...Expression) (Expression, error) {
+				if len(exp) != 1 {
+					return nil, NewRuntimeError("E1007", strconv.Itoa(len(exp)))
+				}
+				if l, ok := exp[0].(*List); ok {
+					if len(l.Value) <= 0 {
+						return nil, NewRuntimeError("E1011", strconv.Itoa(len(l.Value)))
+					}
+					return NewList(l.Value[1:]), nil
+				} else if p, ok := exp[0].(*Pair); ok {
+					return p.Cdr, nil
+				} else {
+					return nil, NewRuntimeError("E1005", reflect.TypeOf(exp[0]).String())
+				}
+			})
+	}
+	buildInFuncTbl["cadr"] = func(exp []Expression, env *SimpleEnv) (Expression, error) {
+		return EvalCalcParam(exp, env,
+			func(exp ...Expression) (Expression, error) {
+				if len(exp) != 1 {
+					return nil, NewRuntimeError("E1007", strconv.Itoa(len(exp)))
+				}
+				if l, ok := exp[0].(*List); ok {
+					if len(l.Value) < 2 {
+						return nil, NewRuntimeError("E1011", strconv.Itoa(len(l.Value)))
+					}
+					return l.Value[1], nil
+				} else {
+					return nil, NewRuntimeError("E1005", reflect.TypeOf(exp[0]).String())
+				}
+			})
+	}
+	buildInFuncTbl["cons"] = func(exp []Expression, env *SimpleEnv) (Expression, error) {
+		return EvalCalcParam(exp, env,
+			func(exp ...Expression) (Expression, error) {
+				if len(exp) != 2 {
+					return nil, NewRuntimeError("E1007", strconv.Itoa(len(exp)))
+				}
+				if _, ok := exp[1].(*List); ok {
+					var args []Expression
+					args = append(args, exp[0])
+					return NewList(append(args, (exp[1].(*List)).Value...)), nil
+				}
+				return NewPair(exp[0], exp[1]), nil
+			})
+	}
+	buildInFuncTbl["append"] = func(exp []Expression, env *SimpleEnv) (Expression, error) {
+		return EvalCalcParam(exp, env,
+			func(exp ...Expression) (Expression, error) {
+				if len(exp) < 2 {
+					return nil, NewRuntimeError("E1007", strconv.Itoa(len(exp)))
+				}
+				var expList []Expression
+				for _, e := range exp {
+					if v, ok := e.(*List); ok {
+						expList = append(expList, v.Value...)
+					} else {
+						return nil, NewRuntimeError("E1005", reflect.TypeOf(e).String())
+					}
+				}
+				return NewList(expList), nil
+			})
+	}
+	buildInFuncTbl["last"] = func(exp []Expression, env *SimpleEnv) (Expression, error) {
+		return EvalCalcParam(exp, env,
+			func(exp ...Expression) (Expression, error) {
+				if len(exp) != 1 {
+					return nil, NewRuntimeError("E1007", strconv.Itoa(len(exp)))
+				}
+				if l, ok := exp[0].(*List); ok {
+					if len(l.Value) <= 0 {
+						return nil, NewRuntimeError("E1011", strconv.Itoa(len(l.Value)))
+					}
+					return l.Value[len(l.Value)-1], nil
+				} else if p, ok := exp[0].(*Pair); ok {
+					return p.Car, nil
+				} else {
+					return nil, NewRuntimeError("E1005", reflect.TypeOf(exp[0]).String())
+				}
+			})
+	}
+	buildInFuncTbl["reverse"] = func(exp []Expression, env *SimpleEnv) (Expression, error) {
+		return EvalCalcParam(exp, env,
+			func(exp ...Expression) (Expression, error) {
+				if len(exp) != 1 {
+					return nil, NewRuntimeError("E1007", strconv.Itoa(len(exp)))
+				}
+				if l, ok := exp[0].(*List); ok {
+					if len(l.Value) <= 1 {
+						return l, nil
+					}
+					args := make([]Expression, len(l.Value))
+					idx := len(l.Value) - 1
+					for _, c := range l.Value {
+						args[idx] = c
+						idx = idx - 1
+					}
+					return NewList(args), nil
+				} else {
+					return nil, NewRuntimeError("E1005", reflect.TypeOf(exp[0]).String())
+				}
+			})
+	}
+	buildInFuncTbl["iota"] = func(exp []Expression, env *SimpleEnv) (Expression, error) {
+		return EvalCalcParam(exp, env,
+			func(exp ...Expression) (Expression, error) {
+				if len(exp) != 1 && len(exp) != 2 {
+					return nil, NewRuntimeError("E1007", strconv.Itoa(len(exp)))
+				}
+				var l []Expression
+				max, ok := exp[0].(*Integer)
+				if !ok {
+					return nil, NewRuntimeError("E1002", reflect.TypeOf(exp[0]).String())
+				}
+				start := 0
+				if len(exp) == 2 {
+					v, ok := exp[1].(*Integer)
+					if !ok {
+						return nil, NewRuntimeError("E1002", reflect.TypeOf(exp[1]).String())
+					}
+					start = v.Value
+				}
+				for i := start; i < start+max.Value; i++ {
+					l = append(l, NewInteger(i))
+				}
+
+				return NewList(l), nil
+			})
+	}
+	buildInFuncTbl["map"] = func(exp []Expression, env *SimpleEnv) (Expression, error) {
+		return EvalCalcParam(exp, env,
+			func(exp ...Expression) (Expression, error) {
+				if len(exp) != 2 {
+					return nil, NewRuntimeError("E1007", strconv.Itoa(len(exp)))
+				}
+				lambda := func(result Expression, item Expression, vaList []Expression) ([]Expression, error) {
+					return append(vaList, result), nil
+				}
+				return listFunc(lambda, exp...)
+			})
+	}
+	buildInFuncTbl["for-each"] = func(exp []Expression, env *SimpleEnv) (Expression, error) {
+		return EvalCalcParam(exp, env,
+			func(exp ...Expression) (Expression, error) {
+				if len(exp) != 2 {
+					return nil, NewRuntimeError("E1007", strconv.Itoa(len(exp)))
+				}
+
+				fn, ok := exp[0].(*Function)
+				if !ok {
+					return nil, NewRuntimeError("E1006", reflect.TypeOf(exp[0]).String())
+				}
+				l, ok := exp[1].(*List)
+				if !ok {
+					return nil, NewRuntimeError("E1005", reflect.TypeOf(exp[1]).String())
+				}
+				param := make([]Expression, 1)
+				for _, param[0] = range l.Value {
+					_, err := fn.Execute(param, nil)
+					if err != nil {
+						return nil, err
+					}
+				}
+				return NewNil(), nil
+			})
+	}
+	buildInFuncTbl["filter"] = func(exp []Expression, env *SimpleEnv) (Expression, error) {
+		return EvalCalcParam(exp, env,
+			func(exp ...Expression) (Expression, error) {
+				if len(exp) != 2 {
+					return nil, NewRuntimeError("E1007", strconv.Itoa(len(exp)))
+				}
+				lambda := func(result Expression, item Expression, vaList []Expression) ([]Expression, error) {
+					b, ok := result.(*Boolean)
+					if !ok {
+						return nil, NewRuntimeError("E1001", reflect.TypeOf(result).String())
+					}
+					if b.Value {
+						return append(vaList, item), nil
+					}
+					return vaList, nil
+				}
+				return listFunc(lambda, exp...)
+			})
+	}
+	buildInFuncTbl["reduce"] = func(exp []Expression, env *SimpleEnv) (Expression, error) {
+		return EvalCalcParam(exp, env,
+			func(exp ...Expression) (Expression, error) {
+				if len(exp) != 3 {
+					return nil, NewRuntimeError("E1007", strconv.Itoa(len(exp)))
+				}
+				fn, ok := exp[0].(*Function)
+				if !ok {
+					return nil, NewRuntimeError("E1006", reflect.TypeOf(exp[0]).String())
+				}
+				l, ok := exp[2].(*List)
+				if !ok {
+					return nil, NewRuntimeError("E1005", reflect.TypeOf(exp[1]).String())
+				}
+				if len(l.Value) == 0 {
+					return exp[1], nil
+				}
+				param := make([]Expression, len(fn.ParamName.Value))
+				result := l.Value[0]
+				for _, c := range l.Value[1:] {
+					param[0] = result
+					param[1] = c
+					r, err := fn.Execute(param, nil)
+					result = r
+					if err != nil {
+						return nil, err
+					}
+				}
+				return result, nil
+			})
+	}
+
+}
